Drop StaticThisUsage and InlineDocCommentDeclaration.InvalidFormat
PHPStan and Psalm both flag `$this` in static methods and malformed inline
`@var` annotations, so CS rules for the same patterns duplicate lint work.
Disable them on both sides:

- `Squiz.Scope.StaticThisUsage` set to severity 0
- `SlevomatCodingStandard.Commenting.InlineDocCommentDeclaration` keeps its
  other codes (InvalidCommentType, MissingVariable, NoAssignment — genuine
  format/structure checks) but excludes InvalidFormat
- `phpdoc_var_annotation_correct_order` set to false so the fixer doesn't
  silently canonicalize the same shape away

Move both fixtures from invalid/ to valid/ and register them in EXPECTED_REL
so the SA-overlap gate asserts both SAs keep flagging them. Rename
this-in-static.php to ThisInStatic.php to satisfy Squiz.Classes.ClassFileName.
