+++ tests/RulesTest.php
@@ -36,7 +36,6 @@ final class RulesTest extends TestCase
         ['bracketed-namespace.php', 'php-cs-fixer'],
         ['deprecated-function.php', 'php-cs-fixer'],
         ['text-before-open-tag.php', 'php-cs-fixer'],
-        ['this-in-static.php', 'php-cs-fixer'],
         ['catch-exception.php', 'php-cs-fixer'],
         ['assignment-in-condition.php', 'php-cs-fixer'],
         ['dead-catch.php', 'php-cs-fixer'],
