--- tests/fixtures/invalid/inline-doc-comment.php
@@ -1,7 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/** @var $foo string */
-$foo = 'bar';
-echo $foo;
