--- tests/fixtures/invalid/this-in-static.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Invalid;
-
-class ThisInStatic
-{
-    public static function foo(): void
-    {
-        echo $this->bar;
-    }
-}
